feat(auth): root-domain login funnel — email -> clinic subdomain redirect
Adds oltigo.com/login funnel so staff no longer need to know their clinic
subdomain. New POST /api/auth/resolve-clinic resolves the clinic(s) for an
email (service-role, tenant-safe: returns only subdomain + name), rate-limited,
never logs PII, and returns an empty list for unknown emails to avoid
enumeration. Login route now server-renders the funnel on the root domain and
the existing credentials form on tenant subdomains (unchanged). i18n in
fr/ar/en/ary; unit tests for all security branches.
