+++ src/locales/ar.json
@@ -169,6 +169,11 @@
   "auth.invalidEmail": "يرجى إدخال عنوان بريد إلكتروني صالح.",
   "auth.invalidOtp": "رمز غير صالح أو منتهي الصلاحية. يرجى المحاولة مرة أخرى.",
   "auth.login": "تسجيل الدخول",
+  "auth.funnelDesc": "أدخل بريدك الإلكتروني المهني للوصول إلى صفحة تسجيل الدخول الخاصة بعيادتك.",
+  "auth.funnelContinue": "متابعة",
+  "auth.funnelSearching": "جارٍ البحث…",
+  "auth.funnelChooseClinic": "اختر عيادتك:",
+  "auth.funnelNoClinic": "لم يتم العثور على عيادة لهذا البريد الإلكتروني. تحقق من العنوان أو تواصل مع مسؤول عيادتك.",
   "auth.loginSubtitle": "سجّل الدخول لإدارة صحتك",
   "auth.mfaBackupCodeLabel": "رمز احتياطي",
   "auth.mfaBackupDesc": "أدخل أحد رموزك الاحتياطية لتسجيل الدخول.",
