+++ src/locales/ary.json
@@ -169,6 +169,11 @@
   "auth.invalidEmail": "يرجى إدخال عنوان بريد إلكتروني صالح.",
   "auth.invalidOtp": "رمز غير صالح أو منتهي الصلاحية. يرجى المحاولة مرة أخرى.",
   "auth.login": "تسجيل الدخول",
+  "auth.funnelDesc": "دخل الإيميل ديال الخدمة باش توصل لصفحة تسجيل الدخول ديال العيادة ديالك.",
+  "auth.funnelContinue": "كمّل",
+  "auth.funnelSearching": "كنقلبو…",
+  "auth.funnelChooseClinic": "ختار العيادة ديالك:",
+  "auth.funnelNoClinic": "ما لقيناش عيادة لهاد الإيميل. تحقق من العنوان ولا تواصل مع مسؤول العيادة ديالك.",
   "auth.loginSubtitle": "سجّل الدخول لإدارة صحتك",
   "auth.mfaBackupCodeLabel": "رمز احتياطي",
   "auth.mfaBackupDesc": "أدخل أحد رموزك الاحتياطية لتسجيل الدخول.",
