+++ src/locales/en.json
@@ -169,6 +169,11 @@
   "auth.invalidEmail": "Please enter a valid email address.",
   "auth.invalidOtp": "Invalid or expired code. Please try again.",
   "auth.login": "Login",
+  "auth.funnelDesc": "Enter your work email to reach your clinic's login.",
+  "auth.funnelContinue": "Continue",
+  "auth.funnelSearching": "Searching…",
+  "auth.funnelChooseClinic": "Choose your clinic:",
+  "auth.funnelNoClinic": "No clinic found for this email. Check the address or contact your clinic administrator.",
   "auth.loginSubtitle": "Sign in to manage your health",
   "auth.mfaBackupCodeLabel": "Backup code",
   "auth.mfaBackupDesc": "Enter one of your backup codes to sign in.",
