+++ src/locales/fr.json
@@ -169,6 +169,11 @@
   "auth.invalidEmail": "Veuillez entrer une adresse e-mail valide.",
   "auth.invalidOtp": "Code invalide ou expiré. Veuillez réessayer.",
   "auth.login": "Connexion",
+  "auth.funnelDesc": "Entrez votre e-mail professionnel pour accéder à la connexion de votre clinique.",
+  "auth.funnelContinue": "Continuer",
+  "auth.funnelSearching": "Recherche…",
+  "auth.funnelChooseClinic": "Choisissez votre clinique :",
+  "auth.funnelNoClinic": "Aucune clinique trouvée pour cet e-mail. Vérifiez l'adresse ou contactez l'administrateur de votre clinique.",
   "auth.loginSubtitle": "Connectez-vous pour gérer votre santé",
   "auth.mfaBackupCodeLabel": "Code de secours",
   "auth.mfaBackupDesc": "Entrez l'un de vos codes de secours pour vous connecter.",
